requirements.txt
selenium=4.0.0
pillow
requests
python-telegram-bot
cryptography==2.8
install --upgrade pip